Bothe WasmSharedWebWorkerConn and WasmSharedWebWorkerMgmtConn can `Connect()` to the web worker
